Projekt

+++ DziennikTreningowy/src/components/IActivity.java
@@ -6,12 +6,12 @@ public interface IActivity {
     String getTitle();
     String getDescription();
     String getDate();
-    float getCalories();
+    int getCalories();
 
     void setId(int id);
     void setType(String type);
     void setTitle(String title);
     void setDescription(String description);
     void setDate(String date);
-    void setCalories(float calories);
+    void setCalories(int calories);
 }
